For fun, not to be merged.

The `tfversion` tests are fairly unique in that they specify `"TF_ACC_TERRAFORM_VERSION"`. As a consequence, each of these tests downloads Terraform into its own temporary directory. This PR asks: what if they shared one temporary directory?

* Find Terraform binaries that we've already downloaded instead of repeatedly downloading them
* Read a `TF_LOG_INSTALLER` environment variable to log hc-install activity. Works best with a branch of hc-install that prepends log levels to logged messages.
* What if ... provider installation can also be de-duplicated?

Before:
```
$ time go test ./tfversion -count=1
ok  	github.com/hashicorp/terraform-plugin-testing/tfversion	223.588s
go test ./tfversion -count=1  45.66s user 29.61s system 33% cpu 3:44.32 total
```

After:
```
$ time go test ./tfversion -count=1
ok  	github.com/hashicorp/terraform-plugin-testing/tfversion	58.465s
go test ./tfversion -count=1  22.14s user 12.61s system 58% cpu 58.982 total
```